# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Berlin MarketHUB API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profiles/{userId}:
    get:
      operationId: getProfile
      tags: [profiles]
      summary: Get a user profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "404":
          description: Not found
    patch:
      operationId: updateProfile
      tags: [profiles]
      summary: Update a user profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
  /profiles/{userId}/follow:
    post:
      operationId: toggleFollowProfile
      tags: [profiles]
      summary: Follow or unfollow a seller
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Follow toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowStatus"
  /profiles/{userId}/stats:
    get:
      operationId: getProfileStats
      tags: [profiles]
      summary: Get seller stats (listings, followers, sales)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Seller stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SellerStats"
  /profiles/{userId}/listings:
    get:
      operationId: fetchSellerListings
      tags: [profiles]
      summary: Get listings by a specific seller
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Seller listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Listing"
...
  /cart:
    get:
      operationId: getCart
      tags: [orders]
      summary: Get current user's cart
      responses:
        "200":
          description: Cart items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    post:
      operationId: addToCart
      tags: [orders]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemInput"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/{itemId}:
    delete:
      operationId: removeFromCart
      tags: [orders]
      summary: Remove item from cart
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders (purchases and sales)
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [purchases, sales]
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: Get notifications for current user
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
        - name: unreadOnly
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/read:
    post:
      operationId: markNotificationsRead
      tags: [notifications]
      summary: Mark notifications as read
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationReadInput"
      responses:
        "200":
          description: Marked as read
components:
  schemas:
...
    Listing:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        images:
          type: array
          items:
            type: string
        category:
          type: string
        condition:
          type: string
        status:
          type: string
          enum: [draft, active, sold, archived, pending_review]
        tags:
          type: array
          items:
            type: string
        shippingFee:
          type: ["number", "null"]
        quantity:
          type: integer
        sellerId:
          type: string
        sellerName:
          type: ["string", "null"]
        sellerAvatar:
          type: ["string", "null"]
        sellerVerified:
          type: boolean
        likesCount:
          type: integer
        viewsCount:
          type: integer
        soldCount:
          type: integer
        rating:
          type: ["number", "null"]
        isLiked:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - title
        - price
        - category
        - status
        - sellerId
        - likesCount
        - viewsCount
        - soldCount
        - isLiked
        - images
        - quantity
        - sellerVerified
        - createdAt
        - updatedAt
...
    Profile:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        fullName:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        websiteUrl:
          type: ["string", "null"]
        facebookUrl:
          type: ["string", "null"]
        instagramUrl:
          type: ["string", "null"]
        twitterUrl:
          type: ["string", "null"]
        isVerified:
          type: boolean
        rating:
          type: ["number", "null"]
        reviewCount:
          type: integer
        followersCount:
          type: integer
        followingCount:
          type: integer
        listingsCount:
          type: integer
        totalSales:
          type: ["number", "null"]
        responseTime:
          type: ["string", "null"]
        joinedAt:
          type: string
        isFollowing:
          type: boolean
        badges:
          type: array
          items:
            $ref: "#/components/schemas/Badge"
      required:
        - id
        - username
        - isVerified
        - reviewCount
        - followersCount
        - followingCount
        - listingsCount
        - isFollowing
        - joinedAt
        - badges
    ProfileUpdate:
      type: object
      properties:
        username:
          type: string
        fullName:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
        country:
          type: string
        location:
          type: string
        websiteUrl:
          type: string
        facebookUrl:
          type: string
        instagramUrl:
          type: string
        twitterUrl:
          type: string
    Badge:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        icon:
          type: string
        earnedAt:
          type: string
      required:
        - id
        - name
        - icon
        - earnedAt
    FollowStatus:
      type: object
      properties:
        following:
          type: boolean
        followersCount:
          type: integer
      required:
        - following
        - followersCount
    SellerStats:
      type: object
      properties:
        totalSales:
          type: number
        itemsSold:
          type: integer
        successRate:
          type: number
        responseTime:
          type: string
        repeatBuyers:
          type: number
        positiveReviews:
          type: integer
        rating:
          type: number
      required:
        - totalSales
        - itemsSold
        - successRate
        - responseTime
        - repeatBuyers
        - positiveReviews
        - rating
...
    Cart:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        subtotal:
          type: number
        shippingTotal:
          type: number
        total:
          type: number
      required:
        - items
        - subtotal
        - shippingTotal
        - total
    CartItem:
      type: object
      properties:
        id:
          type: string
        listingId:
          type: string
        title:
          type: string
        price:
          type: number
        image:
          type: string
        quantity:
          type: integer
        sellerId:
          type: string
        sellerName:
          type: string
      required:
        - id
        - listingId
        - title
        - price
        - image
        - quantity
        - sellerId
        - sellerName
    CartItemInput:
      type: object
      properties:
        listingId:
          type: string
        quantity:
          type: integer
      required:
        - listingId
        - quantity
    Order:
      type: object
      properties:
        id:
          type: string
        listingId:
          type: string
        listingTitle:
          type: string
        listingImage:
          type: string
        price:
          type: number
        buyerId:
          type: string
        sellerId:
          type: string
        buyerName:
          type: ["string", "null"]
        sellerName:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, confirmed, shipped, delivered, cancelled, refunded]
        createdAt:
          type: string
      required:
        - id
        - listingId
        - listingTitle
        - listingImage
        - price
        - buyerId
        - sellerId
        - status
        - createdAt
...
    Notification:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [message, follow, like, order, review, price_drop, promo]
        title:
          type: string
        body:
          type: string
        isRead:
          type: boolean
        data:
          type: object
          additionalProperties: true
        createdAt:
          type: string
      required:
        - id
        - type
        - title
        - body
        - isRead
        - createdAt
    NotificationReadInput:
      type: object
      properties:
        ids:
          type: array
          items:
            type: string
        all:
          type: boolean
      required: []