# lib/api-zod/src/generated/types/notificationReadInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Berlin MarketHUB API specification
 * OpenAPI spec version: 0.1.0
 */

export interface NotificationReadInput {
  ids?: string[];
  all?: boolean;
}