# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Berlin MarketHUB API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: string;
  listingId: string;
  listingTitle: string;
  listingImage: string;
  price: number;
  buyerId: string;
  sellerId: string;
  /** @nullable */
  buyerName?: string | null;
  /** @nullable */
  sellerName?: string | null;
  status: OrderStatus;
  createdAt: string;
}