# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Berlin MarketHUB API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Badge } from './badge';

export interface Profile {
  id: string;
  username: string;
  /** @nullable */
  fullName?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  websiteUrl?: string | null;
  /** @nullable */
  facebookUrl?: string | null;
  /** @nullable */
  instagramUrl?: string | null;
  /** @nullable */
  twitterUrl?: string | null;
  isVerified: boolean;
  /** @nullable */
  rating?: number | null;
  reviewCount: number;
  followersCount: number;
  followingCount: number;
  listingsCount: number;
  /** @nullable */
  totalSales?: number | null;
  /** @nullable */
  responseTime?: string | null;
  joinedAt: string;
  isFollowing: boolean;
  badges: Badge[];
}