# lib/api-zod/src/generated/types/profileUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Berlin MarketHUB API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProfileUpdate {
  username?: string;
  fullName?: string;
  bio?: string;
  avatarUrl?: string;
  country?: string;
  location?: string;
  websiteUrl?: string;
  facebookUrl?: string;
  instagramUrl?: string;
  twitterUrl?: string;
}